fixed metadata error installed sharp for nextjs moved store folder to src dir
