+++ package.json
@@ -46,6 +46,7 @@
     "react-redux": "^8.1.3",
     "sass": "^1.69.2",
     "server-only": "^0.0.1",
+    "sharp": "^0.32.6",
     "typescript": "5.2.2",
     "yup": "^1.3.2"
   },
