+++ src/app/api/reset-password/route.ts
@@ -1,10 +1,11 @@
 import client from "../../../../apollo-client";
 import { NextRequest, NextResponse } from "next/server";
 import { StatusCodes } from "@/constants/status-code";
-import { GET_USER_PASSWORD_CHANGE_REQUEST } from "@/components/store/modules/user-password-change-request/query";
+
 import dayjs from "dayjs";
 import bcrypt from "bcrypt";
-import { UPDATE_USERS } from "@/components/store/modules/user/query";
+import { GET_USER_PASSWORD_CHANGE_REQUEST } from "@/store/modules/user-password-change-request/query";
+import { UPDATE_USERS } from "@/store/modules/user/query";
 
 interface CustomNextApiRequest extends NextRequest {
   json: () => Promise<NForgotPassword.IRequest["body"]>;
