+++ src/store/types/swapi-type.ts
